move CmdGenerateDocumentation to pulsar-proxy for break cycle deps
